Exit fullscreen graceful failure
Resolves #25

User opens fullscreen mode, then something else on the device exits fullscreen mode, and we fail to detect it. Failure to exit fullscreen mode should not create alert and fail. It should gracefully assume that we have made a mistake, and quietly set the app state back to `isFullscreen: false`