Status message should not replace calendar items
Resolves #70

Currently, when the calendar loads (every 5 minutes) a status message appears instead of the events, which all disappear for a moment.

The expected behavior is that the last know events should be displayed the whole time.

Maybe the status message should appear under the clock? (this would place it on the left side below the clock on larger devices)